Improved concurrency of ZContext.
It was using a CopyOnWriteArrayList, that is not a very efficient way of
managing concurrent access. It now uses a concurrent set, created with
ConcurrentHashMap.newKeySet(). So no manual locking is needed.

linger and sndhwm was protected with a lock, I think making them volatile
provides the same semantic (immediat change of value in all threads when set)
with less code.

It slightly changes the semantic of getSockets(). I'm not sure it's very
important.

In isClosed(), a totally useless synchronized(this) is removed.