custom decoder check was invalid
All the check were done using assert, that are disable during runtime,
make the whole method useless. It was printing result to stdout, this is
bad practice. It now use ZError.InstantiationException. The check method
name is more explicit too.
The CustomDecoderTest is updated, and warning removed, because it used
deprecated methods.
It also add a case where ZError.InstantiationException better match the
source exception.