Fixing org.zeromq.ZContext.destroySocket and org.zeromq.ZMQ.Socket.close.
The method destroySocket still set linger, but javadoc says that it should
only be used for special cases. This method is undeprecated, but I'm not
sure what is it's real use case. So deprecation might make sens any way.

The method org.zeromq.ZMQ.Socket.close, when the socket was constructed
from a ZContext, will use a internal org.zeromq.ZContext.closeSocket. This
method is not public because it should not used, only org.zeromq.ZMQ.Socket.close
should be used.