Third try to improve TestZPoller.
It was indeed a problem in the linger of the test. But I also raising problems
with linking of linger in ZMQ.Socket and ZContext. When a socket is obtained
from a ZContext, the linger value of this socket is meaningless. So let
ZContext handle it for all it's owned socket, and explained it in the documentation.

That don't change the old inner working, but make it more explicit in code
and documentation.

Another solution would be to use the context linger only as an hint and let
each socket manage it's own linger, so removing the Socket.setLinger in
ZContext.destroySocket, just using it at socket creation.

Any way, in the test, the linger needs to be set to a non null value.