Finishing improved monitor, with more thorought tests. Some events were missed on client side.
In this branch, if CureServerMechanism refuse a connection because the curve
key is invalid, an hard error (with sending the message ERROR) is send, so
connection is not retried. It's done only on first decryption, because only at this
step it's critical. Latter failures means network error, but that's not sure.
So advices are welcome.