Signaler is now protected from interruption.
It's IO are internal, using a Pipe, so it should never hang. Interrupting
a thread waiting for a Signaler loose events that leaks sockets.

So interruptions are masked in some critical sections, and is transformed
in EINTR error.

Exceptions are not lost any more when closing a Signaler.

A test is added to check this.