Enhance ZMonitor.ZEvent and ZMQ.Event.
ZMonitor.ZEvent can be constructed from a org.zeromq.ZMQ.Event.

Adding an enum for handshake code, as returned from eventHandshakeXXX kind
of events.

Both ZEvent.Event and ZEvent.ProtocolCode can be resolver directly from
the numerical value, to be used in other part of the code.

Adding a value resolver to ZMQ.Event, that resolve the value to an usable
event or integer.

Perhaps a smart way to send a channel in a monitor socket could be created,
instead of discarding it. But it could only be used with inproc://