In zmq.poll.Poller.run(), many exceptions were generating stack dump on stdout. The thread could also silently die, with an out of memory as an example. It's now possible to define two UncaughtExceptionHandler in a Ctx that will be forwarded to the poller.
One (uncaughtExceptionHandler) will be used for the thread's
UncaughtExceptionHandler and should be used to be notify of critical
failures and generate an emergency stop.

The other (nNotificationExceptionHandler) can be used to replace the old
stack dump, for logging or tests.

it can only be used before any socket was created and not modifed any
more.