When doing PLAIN or CURVE authentication, the logic can be exchanged.
The client is doing the bind, and the server doing the connect. That was not handled, a connect socket was not expected to reuse the session, and so found an already configured zapPipe. It’s now handled.

Some error message was not returning any message, the specification says that an empty error should be returned instead.

Big rewrite of the tests for mechanisms, the big test function is splited and more code is shared.