The current version for Google Play is Android 13 (API level 33) as explained at https://support.google.com/googleplay/android-developer/answer/11926878?hl=en.
The Java API that matches this version is Java 11.

But the ABI is still Java 8, for example see the problem with Buffer/ByteBuffer.

Meanwhile, to simplify developement, Google introduced “desugaring”, see
https://developer.android.com/studio/write/java11-default-support-table.

So I think API compliance checking is not needed any more, and removed all
API compliance check that breaks, but not in a meaningfull way.

Of course, support for Java 8 is dropped.